# backend/openapi.yaml
openapi: 3.0.0
paths:
...
  /shanda/quotes/request/{requestId}:
    get:
      operationId: QuotesController_findByRequest
      parameters:
        - name: requestId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QuoteEntity'
      security: &ref_4
        - bearer: []
      summary: Get all quotes for a service request
      tags: &ref_5
        - quotes
  /shanda/quotes:
    post:
      operationId: QuotesController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuoteDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteEntity'
      security: *ref_4
      summary: Create a new quote (active providers only)
      tags: *ref_5
  /shanda/quotes/{id}/accept:
    post:
      operationId: QuotesController_accept
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: ''
      security: *ref_4
      summary: Accept a quote (creates a job)
      tags: *ref_5
  /shanda/quotes/{id}/reject:
    post:
      operationId: QuotesController_reject
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: ''
      security: *ref_4
      summary: Reject a quote
      tags: *ref_5
...
  /shanda/providers:
    get:
      operationId: ProvidersController_findAll
      parameters:
        - name: serviceType
          required: false
          in: query
          schema:
            type: string
        - name: mobileService
          required: false
          in: query
          schema:
            type: boolean
        - name: shopService
          required: false
          in: query
          schema:
            type: boolean
        - name: minRating
          required: false
          in: query
          schema:
            type: number
        - name: limit
          required: false
          in: query
          schema:
            type: number
      responses:
        '200':
          description: ''
      summary: Get all providers with optional filters
      tags: *ref_10
  /shanda/providers/{id}:
    get:
      operationId: ProvidersController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get provider by ID with full profile
      tags: *ref_10
  /shanda/providers/profile:
    put:
      operationId: ProvidersController_updateProfile
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProviderProfileDto'
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: Update provider profile
      tags: *ref_10
  /shanda/providers/onboarding/status:
    get:
      operationId: ProvidersController_getOnboardingStatus
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: Get provider onboarding status and checklist
      tags: *ref_10
  /shanda/providers/onboarding/complete:
    post:
      operationId: ProvidersController_completeOnboarding
      parameters: []
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                type: object
      security:
        - bearer: []
      summary: Complete provider onboarding
      tags: *ref_10
  /shanda/providers/onboarding/start:
    put:
      operationId: ProvidersController_startOnboarding
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
      security:
        - bearer: []
      summary: Start provider onboarding (NONE → DRAFT)
      tags: *ref_10
...
  /shanda/parts:
    post:
      operationId: PartsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePartDto'
      responses:
        '201':
          description: Part created successfully
      security: &ref_16
        - bearer: []
      summary: Add a part to inventory
      tags: &ref_17
        - Parts
  /shanda/parts/my-inventory:
    get:
      operationId: PartsController_getMyInventory
      parameters: []
      responses:
        '200':
          description: ''
      security: *ref_16
      summary: Get provider inventory
      tags: *ref_17
  /shanda/parts/{id}:
    delete:
      operationId: PartsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security: *ref_16
      summary: Remove a part
      tags: *ref_17
...
info:
  title: Shanda Automotive API
  description: Marketplace connecting vehicle owners with service providers
  version: '1.0'
  contact: {}
...
servers:
  - url: http://localhost:4201
    description: Local Development
  - url: https://api.shanda.com
    description: Production
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    QuoteEntity:
      type: object
      properties:
        id:
          type: string
        requestId:
          type: string
        providerId:
          type: string
        amount:
          type: string
        laborCost:
          type: string
          nullable: true
        partsCost:
          type: string
          nullable: true
        estimatedDuration:
          type: string
        description:
          type: string
          nullable: true
        includesWarranty:
          type: boolean
        status:
          type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        provider:
          type: object
      required:
        - id
        - requestId
        - providerId
        - amount
        - estimatedDuration
        - includesWarranty
        - status
        - createdAt
        - updatedAt
    AddQuotePartDto:
      type: object
      properties:
        partId:
          type: string
          format: uuid
          description: Part inventory ID (if from inventory)
        name:
          type: string
          description: Part name
        condition:
          type: string
          enum:
            - OEM
            - AFTERMARKET
            - USED
          description: Part condition
        price:
          type: number
          minimum: 0
          description: Part price
        quantity:
          type: number
          default: 1
          minimum: 1
          description: Quantity
        notes:
          type: string
          description: Additional notes
      required:
        - name
        - condition
        - price
        - quantity
    CreateQuoteDto:
      type: object
      properties:
        requestId:
          type: string
          format: uuid
          description: UUID of the service request
        amount:
          type: string
          pattern: /^\d+(\.\d{1,2})?$/
          description: Total quote amount
          example: '150.00'
        laborCost:
          type: string
          pattern: /^\d+(\.\d{1,2})?$/
          description: Labor cost portion
          example: '100.00'
        partsCost:
          type: string
          pattern: /^\d+(\.\d{1,2})?$/
          description: Parts cost portion
          example: '50.00'
        estimatedDuration:
          type: string
          description: Estimated time to complete
          example: 2 hours
        description:
          type: string
          description: Detailed quote description
        includesWarranty:
          type: boolean
          description: Whether warranty is included
          default: false
        parts:
          description: Parts included in the quote
          type: array
          items:
            $ref: '#/components/schemas/AddQuotePartDto'
      required:
        - requestId
        - amount
        - estimatedDuration
...
    UpdateProviderProfileDto:
      type: object
      properties:
        businessName:
          type: string
          example: Joe's Auto Repair
        serviceTypes:
          example:
            - oil_change
            - brakes
            - tire_service
          type: array
          items:
            type: string
        city:
          type: string
          example: Toronto
        state:
          type: string
          example: 'ON'
        shopCity:
          type: string
          example: Toronto
        shopState:
          type: string
          example: 'ON'
        serviceRadius:
          type: number
          minimum: 1
          example: 25
          description: Service radius in kilometers
...
    CreatePartDto:
      type: object
      properties:
        name:
          type: string
          description: Name of the part
          example: Brake Pads
        category:
          type: string
          description: Part category
          example: brakes
        condition:
          type: string
          description: Condition of the part
          enum:
            - OEM
            - AFTERMARKET
            - USED
          example: AFTERMARKET
        price:
          type: number
          minimum: 0
          description: Price of the part
          example: 49.99
        notes:
          type: string
          description: Additional notes (e.g., fits 2011 Accord 2.4L)
      required:
        - name
        - condition
        - price